refactor: 폴더 계층 구조 변경
