+++ app/logo/controller.go
