+++ app/user/service_test.go
@@ -1,27 +1,27 @@
 package user
 
 import (
-	m "GolangLivePT01/golang_web_programming/membership"
+	"GolangLivePT01/golang_web_programming/app/membership"
 	"fmt"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"testing"
 )
 
-var data = make(map[string]m.Membership)
+var data = make(map[string]membership.Membership)
 
-func createTestMembership(app *m.Service, userName, membershipType string) m.CreateResponse {
-	res, _ := app.Create(m.CreateRequest{UserName: userName, MembershipType: membershipType})
+func createTestMembership(app *membership.Service, userName, membershipType string) membership.CreateResponse {
+	res, _ := app.Create(membership.CreateRequest{UserName: userName, MembershipType: membershipType})
 	return res
 }
 
 func TestLogin(t *testing.T) {
 
 	t.Run("로그인 처리를 합니다.", func(t *testing.T) {
-		app := m.NewService(*m.NewRepository(data))
+		app := membership.NewService(*membership.NewRepository(data))
 		createTestMembership(app, "test", "naver")
 
-		service := NewService(DefaultSecret, *m.NewRepository(data))
+		service := NewService(DefaultSecret, *membership.NewRepository(data))
 		res, err := service.Login("test", "test")
 
 		assert.Nil(t, err)
@@ -30,10 +30,10 @@ func TestLogin(t *testing.T) {
 	})
 
 	t.Run("name 를 입력하지 않았을 때 예외 처리한다.", func(t *testing.T) {
-		app := m.NewService(*m.NewRepository(data))
+		app := membership.NewService(*membership.NewRepository(data))
 		createTestMembership(app, "test", "naver")
 
-		service := NewService(DefaultSecret, *m.NewRepository(data))
+		service := NewService(DefaultSecret, *membership.NewRepository(data))
 		_, err := service.Login("", "test")
 
 		assert.NotNil(t, err)
@@ -41,10 +41,10 @@ func TestLogin(t *testing.T) {
 	})
 
 	t.Run("password 를 입력하지 않았을 때 예외 처리한다.", func(t *testing.T) {
-		app := m.NewService(*m.NewRepository(data))
+		app := membership.NewService(*membership.NewRepository(data))
 		createTestMembership(app, "test", "naver")
 
-		service := NewService(DefaultSecret, *m.NewRepository(data))
+		service := NewService(DefaultSecret, *membership.NewRepository(data))
 		_, err := service.Login("test", "")
 
 		assert.NotNil(t, err)
@@ -52,10 +52,10 @@ func TestLogin(t *testing.T) {
 	})
 
 	t.Run("입력한 id가 존재하지 않을 때 예외 처리한다.", func(t *testing.T) {
-		app := m.NewService(*m.NewRepository(data))
+		app := membership.NewService(*membership.NewRepository(data))
 		createTestMembership(app, "test", "naver")
 
-		service := NewService(DefaultSecret, *m.NewRepository(data))
+		service := NewService(DefaultSecret, *membership.NewRepository(data))
 		_, err := service.Login("test2", "test2")
 
 		assert.NotNil(t, err)
@@ -63,10 +63,10 @@ func TestLogin(t *testing.T) {
 	})
 
 	t.Run("password 가 일치하지 않을 때 예외 처리한다.", func(t *testing.T) {
-		app := m.NewService(*m.NewRepository(data))
+		app := membership.NewService(*membership.NewRepository(data))
 		createTestMembership(app, "test", "naver")
 
-		service := NewService(DefaultSecret, *m.NewRepository(data))
+		service := NewService(DefaultSecret, *membership.NewRepository(data))
 		_, err := service.Login("test", "1234")
 
 		assert.NotNil(t, err)
